Extract binder from mvicore to a separate module
