+++ binder/src/main/java/com/badoo/binder/middleware/config/ConsumerMiddlewareFactory.kt
@@ -1,6 +1,6 @@
-package com.badoo.mvicore.consumer.middlewareconfig
+package com.badoo.binder.middleware.config
 
-import com.badoo.mvicore.consumer.middleware.base.Middleware
+import com.badoo.binder.middleware.base.Middleware
 import io.reactivex.functions.Consumer
 
 typealias ConsumerMiddlewareFactory<T> = (Consumer<T>) -> Middleware<Any, T>
