+++ binder/src/main/java/com/badoo/binder/middleware/config/NonWrappable.kt
@@ -1,0 +1,3 @@
+package com.badoo.binder.middleware.config
+
+interface NonWrappable
