+++ mvicore-android/src/main/java/com/badoo/mvicore/android/AndroidBindings.kt
@@ -1,9 +1,7 @@
 package com.badoo.mvicore.android
 
 import android.arch.lifecycle.LifecycleOwner
-import com.badoo.mvicore.binder.Binder
-import android.arch.lifecycle.Lifecycle as AndroidLifecycle
-import com.badoo.mvicore.binder.lifecycle.Lifecycle as BinderLifecycle
+import com.badoo.binder.Binder
 
 abstract class AndroidBindings<T : Any>(
     lifecycleOwner: LifecycleOwner
