+++ mvicore-android/src/main/java/com/badoo/mvicore/android/lifecycle/StartStopBinderLifecycle.kt
@@ -3,7 +3,7 @@ package com.badoo.mvicore.android.lifecycle
 import android.arch.lifecycle.DefaultLifecycleObserver
 import android.arch.lifecycle.LifecycleOwner
 import android.arch.lifecycle.Lifecycle as AndroidLifecycle
-import com.badoo.mvicore.binder.lifecycle.Lifecycle as BinderLifecycle
+import com.badoo.binder.lifecycle.Lifecycle as BinderLifecycle
 
 
 class StartStopBinderLifecycle(
