+++ mvicore-plugin/middleware/src/main/java/com/badoo/mvicore/middleware/IdeaPluginMiddleware.kt
@@ -1,7 +1,7 @@
 package com.badoo.mvicore.middleware
 
-import com.badoo.mvicore.binder.Connection
-import com.badoo.mvicore.consumer.middleware.base.Middleware
+import com.badoo.binder.Connection
+import com.badoo.binder.middleware.base.Middleware
 import io.reactivex.functions.Consumer
 
 class IdeaPluginMiddleware<Out: Any, In: Any>(
