+++ mvicore-plugin/middleware/src/main/java/com/badoo/mvicore/middleware/data/Connection.kt
@@ -1,8 +1,6 @@
 package com.badoo.mvicore.middleware.data
 
-import com.badoo.mvicore.binder.Connection
-import com.badoo.mvicore.plugin.model.ConnectionData
-import com.badoo.mvicore.plugin.model.Id
+import com.badoo.binder.Connection
 
 fun Connection<out Any, out Any>.parse() =
     ConnectionData(
