+++ mvicore/src/main/java/com/badoo/mvicore/consumer/middleware/ConsumerMiddleware.kt
@@ -1,9 +1,9 @@
 package com.badoo.mvicore.consumer.middleware
 
-import com.badoo.mvicore.consumer.middleware.base.Middleware
+import com.badoo.binder.middleware.base.Middleware
 
 @Deprecated(
     "Left for compatibility reasons",
-    ReplaceWith("Middleware<Any, T>", "com.badoo.mvicore.consumer.middleware.base")
+    ReplaceWith("Middleware<Any, T>", "com.badoo.binder.middleware.base")
 )
 typealias ConsumerMiddleware<T> = Middleware<Any, T>
