--- mvicore/src/main/java/com/badoo/mvicore/consumer/middlewareconfig/NonWrappable.kt
@@ -1,3 +1,0 @@
-package com.badoo.mvicore.consumer.middlewareconfig
-
-interface NonWrappable
