+++ mvicore/src/main/java/com/badoo/mvicore/feature/ReducerFeature.kt
@@ -1,6 +1,6 @@
 package com.badoo.mvicore.feature
 
-import com.badoo.mvicore.consumer.middlewareconfig.NonWrappable
+import com.badoo.binder.middleware.config.NonWrappable
 import com.badoo.mvicore.element.Actor
 import com.badoo.mvicore.element.Bootstrapper
 import com.badoo.mvicore.element.NewsPublisher
