+++ settings.gradle
@@ -1,3 +1,4 @@
+include ':binder'
 include ':mvicore'
 include ':mvicore-diff'
 include ':mvicore-android'
